requirements.txt
# Simplified GitHub Repository Analysis Agent - Dependencies

# --- Web Framework ---
fastapi>=0.110.0
uvicorn>=0.29.0

# --- GitHub Integration ---
PyGithub>=2.3.0

# --- LLM Provider (Groq) ---
langchain-groq>=0.1.0
langchain-core>=0.2.0

# --- Environment & Config ---
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# --- RAG Vector Store ---
chromadb>=0.4.0
sentence-transformers>=2.2.0
